pytaco/pyproject.toml
[tool.poetry]
name = "pytaco"
version = "0.1.0"
description = "A Framework for creating conversational agents using stories and intents"
authors = ["vishnuvyas sethumadhavan <vishnuvyas@gmail.com>"]
license = "GPL v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.7"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"
fastapi = {extras = ["all"], version = "^0.95.2"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
typeguard = "^4.0.0"
mypy = "^1.3.0"
pytest-mypy = "^0.10.3"
ipython = "^8.13.2"
pytest-asyncio = "^0.21.0"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --typeguard-packages=taco --mypy"
testpaths = ["tests"]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
